// client/src/utils/types.ts
export interface SignupData {
  email: string;
  fullName: string;
  password: string;
  cPassword: string;
  showP: boolean;
  showCP: boolean;
  imageUrl: string;
}